+++ tests/testthat/test-mkdirs.R
@@ -1,8 +1,0 @@
-context("mkdirs")
-
-## TODO: Rename context
-## TODO: Add more tests
-
-test_that("we can create a directory", {
-  expect_true(adls_mkdirs())
-})