+++ NEWS.md
@@ -1,5 +1,7 @@
 # AzureDatalakeStore 0.0.0.9000
 
+* Adds `adls_mkdirs()` function to make a directory.
+
 * Adds `adls` S3 class to contain base-URL and token, uses `adls()` as constructor.
 
 * Adds function `adls_url()` to compose the base URL.